Write Tests for categories controller. Change this.User.Identity.Name with this.User.Identity.IsAuthenticated

+++ RestaurantsPlatform/Web/RestaurantsPlatform.Web/Areas/Categories/Controllers/CategoriesController.cs
@@ -163,6 +163,11 @@
         [HttpPost]
         public async Task<IActionResult> Delete(DeleteCategoryInputModel input)
         {
+            if (!this.ModelState.IsValid)
+            {
+                return this.View(input);
+            }
+
             await this.categoryService.DeleteCategoryAsync(input.Id);
 
             this.TempData[SuccessNotification] = SuccessfullyDeletedCategory;
